+++ src/main/data/data.txt
@@ -1,2 +1,4 @@
-T | 0 | sleep
+T | 1 | sleep
 E | 0 | artbox | 4/1/6969 1420 | 31/12/9999 2359
+T | 0 | gym
+D | 0 | submission | 12/12/2001 1600
